updated evaluate gender and evaluate interval and moved them into separate modules
